+++ R/plot_GSnetwork.R
@@ -1,25 +1,26 @@
-#' @title Plot significantly perturbed gene-sets as network
+#' @title Plot significantly perturbed gene-sets as a network
 #'
-#' @param normalisedScores A dataframe derived from the `normalise_by_permu()` function
+#' @param normalisedScores A `data.frame` of pathway perturbation scores derived from the `normalise_by_permu()` function
 #' @param gsTopology List of pathway topology matrices generated using function `retrieve_topology()`
-#' @param colorBy Choose to color nodes either by *robustZ* or *pvalue* A column must exist in the normalisedScores for the chosen parameter
-#' @param foldGSname `logical`. Should long gene-set names be folded into two lines
+#' @param colorBy Choose to color nodes either by *robustZ* or *pvalue*. A column must exist in the `normalisedScores`
+#' `data.frame` for the chosen parameter
+#' @param foldGSname `logical` Should long gene-set names be folded into two lines
 #' @param foldafter The number of words after which gene-set names should be folded. Defaulted to 2
 #' @param layout The layout algorithm to apply. Accept all layout supported by `igraph`
-#' @param edgeAlpha Transparency of edges. Default to 0.8
+#' @param edgeAlpha `numerical` Transparency of edges. Default to 0.8
 #' @param up_col The color used to label activated gene-sets. Only applicable if `colorBy` is set to be "robustZ"
 #' @param down_col The color used to label inhibited gene-sets. Only applicable if `colorBy` is set to be "robustZ"
 #' @param scale_edgeWidth A numerical vector of length 2 to be provided to `ggraph::scale_edge_width_continuous()` for specifying
-#' the minimum and maximum edge widths after transformation. Defaulted to c(0.5, 3)
+#' the minimum and maximum edge widths after transformation. Defaulted to `c(0.5, 3)`
 #' @param edgeLegend logical` Should edge weight legend be shown
 #' @param scale_nodeSize A numerical vector of length 2 to be provided to `ggplot2::scale_size()` for specifying
-#' the minimum and maximum node sizes after transformation. Defaulted to c(3,6)
-#' @param nodeShape The shape to use for nodes
+#' the minimum and maximum node sizes after transformation. Defaulted to `c(3,6)`
+#' @param nodeShape Shape of nodes
 #' @param color_lg `logical` Should color legend be shown
-#' @param color_lg_title Title for the color legend
+#' @param color_lg_title Optional. Title for the color legend
 #' @param lb_size Size of node text labels
 #' @param lb_color Color of node text labels
-#' @param plotIsolated `logical`.Should nodes not connected to any other nodes be plotted.  Default to FALSE
+#' @param plotIsolated `logical` Should nodes not connected to any other nodes be plotted.  Default to FALSE
 #' @importFrom ggraph ggraph geom_node_point geom_node_text geom_edge_link
 #' @importFrom ggplot2 scale_color_continuous scale_color_manual aes_ guides aes scale_size
 #' @importFrom ggraph scale_edge_width_continuous
@@ -29,8 +30,10 @@
 #' @examples
 #' load(system.file("extdata", "gsTopology.rda", package = "sSNAPPY"))
 #' load(system.file("extdata", "normalisedScores.rda", package = "sSNAPPY"))
+#'
 #' #Subset pathways significantly perturbed in sample R5020_N2_48
 #' subset <- dplyr::filter(normalisedScores, adjPvalue < 0.05, sample == "R5020_N2_48")
+#'
 #' # Color network plot nodes by robust z-score
 #' plot_gs_network(subset, gsTopology,
 #' colorBy = "robustZ", layout = "dh",
@@ -51,6 +54,7 @@ plot_gs_network <- function(normalisedScores, gsTopology, colorBy = c("robustZ",
     # Make sure the gs topologies are a named list with at least two elements
     stopifnot(length(names(gsTopology)) == length(gsTopology))
     if(length(unique(normalisedScores$gs_name)) < 2) stop("At least 2 gene-sets are required for a network plot")
+    gsTopology <- gsTopology[names(gsTopology) %in% normalisedScores$gs_name]
 
     # create igraph object
     g <- make_gsNetwork(normalisedScores, gsTopology, colorBy = colorBy, plotIsolated = plotIsolated)
@@ -86,13 +90,13 @@ plot_gs_network <- function(normalisedScores, gsTopology, colorBy = c("robustZ",
 }
 
 #' @importFrom reshape2 melt
+#' @importFrom dplyr bind_rows
 #' @importFrom igraph E V graph.data.frame set_edge_attr set_vertex_attr degree delete_vertices delete_edges
 make_gsNetwork <- function(normalisedScores, gsTopology,  colorBy = c("robustZ", "pvalue", "community"),
                            plotIsolated ){
 
     # create dummy variable to pass R CMD CHECK
     from <- to <- E <- robustZ <- NULL
-    gsTopology <- gsTopology[names(gsTopology) %in% normalisedScores$gs_name]
     GS2Gene <- get_GSgenelist(gsTopology)
     GS2Gene <- left_join(normalisedScores, GS2Gene, by = "gs_name")
 
@@ -147,10 +151,10 @@ get_GSgenelist <- function(gsTopology, mapEntrezID = NULL){
     colnames(GStoGene) <- c("entrezid", "gs_name")
 
     if (all(c("entrezid","mapTo") %in% colnames(mapEntrezID))){
-        if (all(GStoGene$entrezid %in% mapEntrezID$entrezid)){
+        if (any(GStoGene$entrezid %in% mapEntrezID$entrezid)){
             left_join(GStoGene, mapEntrezID[,c("entrezid","mapTo")], by = "entrezid")
         } else {
-            warning("Not all entrez IDs were mapped. Entrez IDs were plotted.")
+            warning("None of the Entrez IDs in mapEntrezID mapped to gsTopology.")
             return(GStoGene)
 
         }